refactory/#87 User->Character 네이밍 변경
